requirements.txt
# Framework para a API web
fastapi

# Servidor para rodar a API
uvicorn[standard]

# Bibliotecas para o web scraping
requests
beautifulsoup4

# Para JWT (tokens) e criptografia
python-jose[cryptography]

# Para hash (criptografia) de senhas
passlib[bcrypt]

# Necessário para o formulário de login do FastAPI (OAuth2PasswordRequestForm)
python-multipart

# Para carregar variáveis de ambiente do arquivo .env
python-dotenv

# ORM para interagir com o banco de dados
SQLAlchemy

# Driver para conectar o SQLAlchemy ao MySQL
PyMySQL
